parse.pl Changed: Production names are now abbreviated in all messages.
In other words, the package names are always omitted.
This makes messages more compact and easier to read.

To avoid having to remove the package names every time that a message is
composed, the shortened name is stored in a new field in the production,
called `PROD_SHORT_NAME`.
Note that the `unshift` call in function `def` needed to be changed.